+++ Modules/Microsoft365DSC/Examples/Resources/EXODkimSigningConfig/1-Create.ps1
@@ -12,12 +12,13 @@ Configuration Example
     )
     Import-DscResource -ModuleName Microsoft365DSC
 
+    $Domain = $Credscredential.Username.Split('@')[1]
     node localhost
     {
         EXODkimSigningConfig 'ConfigureDKIMSigning'
         {
             KeySize                = 1024
-            Identity               = 'contoso.onmicrosoft.com'
+            Identity               = $Domain
             HeaderCanonicalization = "Relaxed"
             Enabled                = $True
             BodyCanonicalization   = "Relaxed"