+++ Modules/Microsoft365DSC/Examples/Resources/EXORemoteDomain/1-Create.ps1
@@ -29,7 +29,7 @@ Configuration Example
             IsInternal                           = $False
             LineWrapSize                         = "Unlimited"
             MeetingForwardNotificationEnabled    = $False
-            Name                                 = "Default"
+            Name                                 = "Integration"
             NonMimeCharacterSet                  = "iso-8859-1"
             PreferredInternetCodePageForShiftJis = "Undefined"
             TargetDeliveryDomain                 = $False