+++ CHANGELOG.md
@@ -17,6 +17,13 @@
     FIXES [#4256](https://github.com/microsoft/Microsoft365DSC/issues/4256)
 * TeamsFilesPolicy
   * Add default value ('Present') to parameter Ensure
+* TeamsEmergencyCallRoutingPolicy
+  * Fix deletion of resource
+    FIXES [#4261](https://github.com/microsoft/Microsoft365DSC/issues/4261)
+* TEAMS
+  * Added support for ManagedIdentity Authentication across Teams resources.
+* DEPENDENCIES
+  * Updated MSCloudLoginAssistant dependencies to version 1.1.10.
 
 # 1.24.124.1